Server architecture skeleton
This begins some of the core classes required for the websocket server
architecture, including the management of connections and active raids.
Add basic handling for two client events: starting and completing a
raid.

Additionally, extracts raid event definitions to a common/ module shared
by both the server and web app.