Connect web app to database
This adds a mongoose driver to the main web app which connects to the
same database as the server, and updates the raids page to pull data
from it instead of a JSON file.

Additionally, updates the development mongo Docker image to populate the
database with an exported raid dump on first initialization.