Fix database seed data
The version of the plugin which generated the old data set had a bug
causing it to incorrectly report NPC updates. This replaces the old seed
data with a new set of raids.

The new data contains PLAYER_ATTACK events, so this also defines those
events in the database schema.