Start adding NPC definitions, fix tick control
This defines an NpcId enum which will hold all known blert NPCs, and
mapping of IDs to NPC definitions including information such as their
name and size.

Additionally, corrects a bug where the parsed tick was being clamped to
NaN as the raid data had not yet been initialized, and updates the tick
input field to allow users to clear it.