Added splits to attack timeline (#11)
Splits now appear on the attack timeline with parameters: `tick`
(number), `splitName` (string) and `splitCustomContent` (arbitrary
HTML). Visually they appear as stylized "brackets" above and below a
column on the timeline. Clicking on one of the split titles will open a
tooltip with the `splitCustomContent` as its contents.

Also added the <LigmaTooltip /> component for rendering tooltips. Uses
React Portal to render the tooltip outside of the normal DOM hierarchy
and the library react-tooltip (also installed as a part of this repo) to
handle the positioning and visibility of the tooltip.