Read items names from a cache instead of the DB
Adds a JSON dump of all item names in the game and a cache for looking
up names by ID. Updates the room event fetcher to populate item names
from this cache instead of reading from the database.

Removing item names from the database results in a ~20% reduction in
storage and socket payload sizes. This could be further extended to web
API payloads in the future, if the cache is provided to the web client.