Feature Addition: Tick Timeline Scrubber
This PR adds a timeline scrubber to the tick timeline. This allows users to see a birds eye view of the attack timeline and to scroll to whatever key splits they want to see. Key splits are shown as tooltips above and below the timeline (alternating). Clicking on a split will scroll the timeline to that split.

Also passes splits to the controls component on each boss page, and sets-up on a per-boss basis what each boss room considers an "unimportant" split.