Migrate websocket wire format to protobuf (#37)
This change transitions the serialization format of events sent over the
blert websocket from JSON to protobuf. As part of this, several
coinciding changes are made to the database document schemas. Most
notably, all enum values, which were previously serialized as strings in
the JSON payload, are converted to integers. Alongside this, several
other fields are renamed.