Implement backend for name changes
This adds a system which updates a player's in-game username to a new
one, migrating any data collected under the new name since the last
challenged recorded under the old name over to the same player.

This only implements the backend side; no user-facing API to request a
name change yet exists.

To facilitate testing name changes, this also sets up Jest as a unit
test runner and integrates it to the build.