Track active challenges by OSRS username
This creates a mapping of OSRS usernames to challenges in which they are
participating. When a client logs into the game, looks up a challenge
for their username, and if one exists, requests the client to confirm
they are still in it.

This allows clients to reconnect following a hard disconnection and
either rejoin their previous challenge, or notify the server that it has
ended in without a regular CHALLENGE_END event.