Track accuracy of individual room events
This adds a flag to every room event marking whether it is accurate or
not, to allow excluding bad data from large aggregate queries of events.
The flag is set based on the recorded room timer when a stage is
completed, as well as whether events were lost due to a primary client
switch.