Migrate event server to Postgres and file backends
This replaces all instances of MongoDB in the event server with a
combination of Postgres and static file data. User and player management
is rewritten in terms of SQL queries using the new database schemas,
and most challenge data is written to protobuf-encoded files.

The API for interacting with static protobuf files is generalized and
extracted from the web app into the shared common module.