Relax accuracy check on challenge completion times
The final times recorded at the end of a challenge were only labeled as
accurate if every stage time within that challenge was also accurate.
This is unnecessarily strict, as each individual stage time is confirmed
on stage completion, and therefore its accuracy is only relevant to
events occurring within the stage.

This removes that requirement, allowing overall times to be accurate as
long as every room is completed (so its time is known to have been
confirmed).