Ignore events for unknown players
There are rare cases in which a challenge's party may not be properly
intialized. This safeguards against that by ignoring the event for the
invalid player, instead of erroring and causing every event in the
payload the be dropped. A log is added to aid with identifying the
invalid player to find the root cause of the party problem.