Begin remote challenge processing
This begins adding challenge processing to the challenge server by
setting up a manager which stores active challenge state in a Redis
database, and implements the initial necessary infrastructure to start
and end challenges.

On the event server, a RemoteChallengeManager implementation is started,
which makes calls to a configured challenge server to handle challenge
start and end events.