Handle stage updates from a single user
This adds stage update support to the challenge server and remote
challenge processor. Starting a new stage updates the state of the
challenge in Redis. When a stage is completed, the stage information
(and future events) from its clients are pulled into a Merger, whose
output is then fed to the challenge processor to update the challenge
state in the database and save challenge data to the data repository.

Only a single challenge client is supported initial; multiple clients
will be handled in a future commit.