Process challenge events at the end of a stage
This updates the event server to write received challenge events to a
Redis stream. Once a stage is complete, the challenge server pulls
events from this stream and forwards them to a ChallengeProcessor.

Code from the original Challenge class in the event server is ported
over to the ChallengeProcessor and its implementations to process all
challenge events for a stage at once.