Set stage end timeout following first end event
Once the first connected client in a challenge indicates a stage has
finished, if there are other clients, set a timeout to allow their
stage end events to arrive. Process the stage end either when the final
client's event is received, or once the timeout expires if not every
client notifies the server.