Handle client activity state changes
This implements the setClient{Active,Inactive} operations on the
challenge server in response to remote client events. Activity state
is used to control challenge cleanup timeouts. Additionally, on
completion of a stage end timeout, checks whether any clients are
still active in the challenge, and starts a cleanup timeout if not.