Handle client reconnection
This implements the last parts of the ChallengeManager API on both the
challenge server and RemoteChallengeManager: querying an active
challenge, and re-adding a client to an existing challenge.
Additionally, moves the player to challenge mapping of the PlayerManager
into Redis. Combined, these changes allow the event server to confirm a
client's challenge state when they become active, and either rejoin or
remove them from an existing challenge.