Improve total stage deaths query
Stage deaths were being counted based on PLAYER_DEATH events in the
database. However, the database only contains events from stages whose
time is verified to be accurate, which means in practice that events
from stages which were not completed are not saved. Because uncompleted
(e.g. wiped) stages are a significant source of deaths, the database
stage numbers are severely underreported.

As stage time accuracy doesn't affect whether or not a death actually
happened, there is no reason to limit deaths to those occurring in
accurate stages. The database already has another source of aggregate
death data: individual player stats, which are updated regardless of
stage accuracy.

This change updates the total stage deaths aggregation to use player
stats data for stages where this is available, while falling back to
the existing method for stages not tracked through stats.