Prevent left nav open when touching scrollable
If the user is touching an element that is horizontally scrollable, they
are probably trying to scroll it and not open the menu. This updates the
menu's touch listener to check for these cases and only register a menu
open if the touch occurs near the left edge of the screen.