Initial setup viewing page and voting system
Creates a setup viewer which renders the setup with its metadata,
revision history, and a new item counts component displaying how many of
each item exists in the setup.

Additionally, creates a database table for tracking users' votes on
setups and implements adding and removing votes on setups in the viewer.