Support exporting and importing inventory setups
This implements a translation module which converts Blert setups to
other common setup formats. The first format implemented is the one
used by the inventory setups plugin. The export button on a setup view
page is given a menu where users can select a format and player to
export, with the result being copied to their clipboard.