Redesign challenge layout and overview pages
Changes the layouts of the ToB and Colosseum challenge pages to use a
top nav bar to navigate between stages in the challenge, instead of the
previous submenus in the sidebar, which were unintuititive.

Additionally, redesigns the overview pages for both these challenge
types to be in line with the new site UI design guidelines.