Redesign boss pages
THis applies the site UI redesign to every boss page (ToB bosses,
Colosseum waves), bringing them in line with the new styling of the rest
of the site. As part of this, it removes the old tabbed mobile view,
replacing it with a single scrolling page without any custom compact
display code.