Fix heatmap challenge count on player overview
The player challenge heatmap shows how many challenges the player
completed each day over the previous year. There was a bug with the
data fetching query which did not set a start date, causing it to
return aggregate data since the player was first recorded. This
corrects the query to only aggregate challenges starting at most a
year ago.