Refactor challenge creation
Previously, the entire challenge creation process including database
writes was performed within a Redis transaction. This resulted in a
race condition where multiple concurrent clients in the same challenge
could create a new database record, only one of which would be used,
leaving the other in a "zombie" state.

This separates the logic to decide which client creates a challenge
from its creation by splitting the process into two steps:

1. A Redis transaction is used to decide which client is responsible
   for creating a challenge. The selected client writes only minimal
   state to Redis.
2. Following this transaction, the selected client proceeds to create
   the challenge in the database. Any other clients that concurrently
   join block until the record has been created and the challenge is
   considered active.

Doing so ensures that only one record can be created for a challenge
and makes the process more robust in general.

Additionally, this also adds a `finishTime` to challenge records,
indicating the real world time at which they complete.