Clean up sessions without challenges
Sessions are created whenever a new team starts a challenge. If that
challenge is deleted (e.g due to ending without data), the session
would become an orphan. This updates session finalization to check
how many challenges a session holds when it completes. If there are
none, the session is deleted. Additionally, if every challenge within
the session has been abandoned, the session is marked with a new HIDDEN
status, as abandoned challenges are omitted from results by default.