Redesign challenge pages to show session data
This updates the challenge pages (`/raids/tob`, `/challenges/colosseum`
etc.) to display session information.

The list of recent challenges is replaced by recent sessions, each
showing overall session information and allowing expansion to list
individual challenges within the session.

The stats section of the page is redone to include additional session
statistics, as well as displaying the most active team for the day.

The core challenge page is created as a reusable component, allowing
challenge page.tsx files to be minimal in size.