Properly implement stage attempts
The implementation of stage attempts adedd alongside the Mokhaiotl
challege was naive and did not work properly. It failed to account
for stage attempts in most of the Redis data structures used by the
challenge server to manage challenge and client state.

This redoes stage attempts to be handled at all parts of the challenge
lifecycle, including support for multiple clients retrying stages of
a challenge. As part of this, a significant API change is made to the
challenge server, where all of its management endpoints now return the
current state of the challenge instead of an empty response.