Update client classification in merger
Changes client classification to improve how it selects a base client,
and to separately choose a reference tick count. Classification now
uses majority vote when looking at server tick counts instead of
assuming that all clients are valid. Additionally, more aggressively
demotes clients which claim to be accurate when tick counts differ.
Changes the definition of matching clients to require accuracy.

This is done as an initial step to implementing the complete merging
algorithm.