merging: Create SimilarityScorer to compare ticks
Creates a scorer which looks at the state of a stage and the events on
two recorded ticks and returns a numeric score of the likelihood that
they represent the same game server tick.

The scorer first checks for basic compatibility between the ticks: all
mutually visible actors must be the in the exact same positions, and
players must have the same visible gear equipped. Any discrepancies
indicate that the two cannot be the same tick.

If the check passes, scores the ticks based on events occurring in them:
attacks, deaths, fuzzy NPC hitpoints, etc. Each component contributes a
different amount to the overall score. Weights for each are initial
guesses and will require proper tuning over time.

Not all scoring components are implemented immediately. Notable,
stage-specific events are omitted from the initial version.