Explicitly request game state from clients
On connecting, the socket server sends a CONNECTION_RESPONSE message to
the connected client. Previously, the client would follow up by sending
a GAME_STATE update after receiving the response if they were logged in
to notify the server of the logged-in account.

This caused a race condition on the plugin side under certain
circumstances: for example, if the socket connection was automatically
reopened when the player was logging in, the CONNECTION_RESPONSE could
arrive before the player had their full account information available,
preventing the GAME_STATE from being sent.

This changes the game state request flow to be explicit and driven by
the server. When a new client connects, the server sends the same
CONNECTION_RESPONSE as before, but then follows it up with a
GAME_STATE_REQUEST. This initial request is sent after a slight delay
to allow for account info to become available in the login case. If the
client does not respond, the server retries the GAME_STATE_REQUEST up to
5 times in total.

This is backwards compatible with older clients, as they will just send
their GAME_STATE update (or not) on the initial CONNECTION_RESPONSE and
ignore GAME_STATE_REQUEST packets.

In addition to the game state changes, this also updates the server to
process the account hash field sent by newer clients in a logged-in
GAME_STATE update. The hash is stored in the players table, and used for
identity checks instead of username going forward.