Persistent user settings
Creates a system for components to define custom persistent settings
which are stored in the database for the active user or in localStorage
when logged out.

Each setting has a string key and an arbitrary JSON value which are
stored as jsonb in a new user_settings table. Settings from localStorage
are synced to the database when a user logs in.