Dynamically send attack definitions to clients
Updates the socket server to manage its own repository of attack
definitions and transmit them to connected plugins, allowing for
addition of attacks without a plugin update.

Creates an admin endpoint for uploading a new list of attacks via a
standardized JSON schema and broadcasting the update to all connected
clients.