Refactor attack timeline tooltip
Extracts the attack timeline tooltip into its own file and modifies how
it works. Instead of passing information in through data attributes,
updates it to use context to access the full player and NPC states. Only
a minimal set of data attributes to indentify which actor the tooltip is
for are kept.